requirements.txt
sqlalchemy
psycopg2-binary
python-dotenv
asyncpg
python-telegram-bot
python-telegram-bot[ext]
python-dotenv
pydantic